Move static check from common to isv
The checks are usually skipped by support to it doesn't make sense to
have them in isv pipelines.
